backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
minio
yt-dlp
openai-whisper
requests
python-multipart
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
